Update the project modules
