+++ src/modules/clone.js
@@ -1,0 +1,1 @@
+export const cloneCollection = (collection) => [...collection];
