+++ src/style.css
@@ -19,7 +19,7 @@ ul {
   list-style: none;
 }
 
-svg {
+img {
   pointer-events: none;
 }
